split utils.ts in smaller files, move utilities only used in one place to that place.
