+++ packages/runner/src/doc.ts
@@ -20,8 +20,10 @@ import { type EntityId } from "./doc-map.ts";
 import type { IRuntime } from "./runtime.ts";
 import { type ReactivityLog } from "./scheduler.ts";
 import { type Cancel } from "./cancel.ts";
-import { arrayEqual } from "./utils.ts";
 import { Labels } from "./storage.ts";
+import { arrayEqual } from "./type-utils.ts";
+
+// Remove the arrayEqual function since we import it now
 
 /**
  * Lowest level cell implementation.
