+++ packages/runner/src/index.ts
@@ -27,9 +27,7 @@ export { effect } from "./reactivity.ts";
 export { type AddCancel, type Cancel, noOp, useCancelGroup } from "./cancel.ts";
 export { Storage } from "./storage.ts";
 export { type ConsoleEvent, ConsoleMethod } from "./harness/console.ts";
-export {
-  addCommonIDfromObjectID,
-  followAliases,
-  maybeGetCellLink,
-} from "./utils.ts";
+export { addCommonIDfromObjectID } from "./data-updating.ts";
+export { followAliases } from "./link-resolution.ts";
+export { maybeGetCellLink } from "./link-resolution.ts";
 export * from "./recipe-manager.ts";
