+++ packages/runner/src/runner.ts
@@ -1,11 +1,14 @@
-import { isRecord } from "@commontools/utils/types";
+import { isObject, isRecord, type Mutable } from "@commontools/utils/types";
 import {
   type Alias,
   isAlias,
   isModule,
   isRecipe,
+  isStatic,
   isStreamAlias,
+  type JSONSchema,
   type JSONValue,
+  markAsStatic,
   type Module,
   type NodeFactory,
   popFrame,
@@ -20,30 +23,22 @@ import {
 import { type DocImpl, isDoc } from "./doc.ts";
 import { type Cell } from "./cell.ts";
 import { type Action, type ReactivityLog } from "./scheduler.ts";
+import { containsOpaqueRef, deepCopy } from "./type-utils.ts";
+import { diffAndUpdate } from "./data-updating.ts";
 import {
-  containsOpaqueRef,
-  deepCopy,
-  diffAndUpdate,
-  extractDefaultValues,
   findAllAliasedCells,
-  followAliases,
-  maybeGetCellLink,
-  mergeObjects,
-  sendValueToBinding,
   unsafe_noteParentOnRecipes,
   unwrapOneLevelAndBindtoDoc,
-} from "./utils.ts";
+} from "./recipe-binding.ts";
+import { followAliases } from "./link-resolution.ts";
+import { maybeGetCellLink } from "./link-resolution.ts";
+import { sendValueToBinding } from "./recipe-binding.ts";
 import { type AddCancel, type Cancel, useCancelGroup } from "./cancel.ts";
 import "./builtins/index.ts";
 import { type CellLink, isCell, isCellLink } from "./cell.ts";
 import { isQueryResultForDereferencing } from "./query-result-proxy.ts";
 import { getCellLinkOrThrow } from "./query-result-proxy.ts";
 import type { IRunner, IRuntime } from "./runtime.ts";
-
-const moduleWrappers = {
-  handler: (fn: (event: any, ...props: any[]) => any) => (props: any) =>
-    fn.bind(props)(props.$event, props),
-};
 
 export class Runner implements IRunner {
   readonly cancels = new WeakMap<DocImpl<any>, Cancel>();
@@ -761,3 +756,87 @@ export class Runner implements IRunner {
     addCancel(this.cancels.get(resultCell.sourceCell!));
   }
 }
+
+/**
+ * Extracts default values from a JSON schema object.
+ * @param schema - The JSON schema to extract defaults from
+ * @returns An object containing the default values, or undefined if none found
+ */
+export function extractDefaultValues(
+  schema: JSONSchema,
+): JSONValue | undefined {
+  if (typeof schema !== "object" || schema === null) return undefined;
+
+  if (
+    schema.type === "object" && schema.properties && isObject(schema.properties)
+  ) {
+    // Ignore the schema.default if it's not an object, since it's not a valid
+    // default value for an object.
+    const obj = deepCopy(isRecord(schema.default) ? schema.default : {});
+    for (const [propKey, propSchema] of Object.entries(schema.properties)) {
+      const value = extractDefaultValues(propSchema);
+      if (value !== undefined) {
+        (obj as Record<string, unknown>)[propKey] = value;
+      }
+    }
+
+    return Object.entries(obj).length > 0 ? obj : undefined;
+  }
+
+  return schema.default;
+}
+
+/**
+ * Merges objects into a single object, preferring values from later objects.
+ * Recursively calls itself for nested objects, passing on any objects that
+ * matching properties.
+ * @param objects - Objects to merge
+ * @returns A merged object, or undefined if no objects provided
+ */
+export function mergeObjects<T>(
+  ...objects: (Partial<T> | undefined)[]
+): T {
+  objects = objects.filter((obj) => obj !== undefined);
+  if (objects.length === 0) return {} as T;
+  if (objects.length === 1) return objects[0] as T;
+
+  const seen = new Set<PropertyKey>();
+  const result: Record<string, unknown> = {};
+
+  for (const obj of objects) {
+    // If we have a literal value, return it. Same for arrays, since we wouldn't
+    // know how to merge them. Note that earlier objects take precedence, so if
+    // an earlier was e.g. an object, we'll return that instead of the literal.
+    if (
+      typeof obj !== "object" ||
+      obj === null ||
+      Array.isArray(obj) ||
+      isAlias(obj) ||
+      isCellLink(obj) ||
+      isDoc(obj) ||
+      isCell(obj) ||
+      isStatic(obj)
+    ) {
+      return obj as T;
+    }
+
+    // Then merge objects, only passing those on that have any values.
+    for (const key of Object.keys(obj)) {
+      if (seen.has(key)) continue;
+      seen.add(key);
+      const merged = mergeObjects<T[keyof T]>(
+        ...objects.map((obj) =>
+          (obj as Record<string, unknown>)?.[key] as T[keyof T]
+        ),
+      );
+      if (merged !== undefined) result[key] = merged;
+    }
+  }
+
+  return result as T;
+}
+
+const moduleWrappers = {
+  handler: (fn: (event: any, ...props: any[]) => any) => (props: any) =>
+    fn.bind(props)(props.$event, props),
+};
