+++ packages/runner/test/utils.test.ts
@@ -1,19 +1,11 @@
 import { afterEach, beforeEach, describe, it } from "@std/testing/bdd";
 import { expect } from "@std/expect";
 import { ID, ID_FIELD } from "@commontools/builder";
-import {
-  addCommonIDfromObjectID,
-  applyChangeSet,
-  diffAndUpdate,
-  extractDefaultValues,
-  followAliases,
-  isEqualCellLink,
-  mergeObjects,
-  normalizeAndDiff,
-  sendValueToBinding,
-  setNestedValue,
-  unwrapOneLevelAndBindtoDoc,
-} from "../src/utils.ts";
+import { addCommonIDfromObjectID, applyChangeSet, diffAndUpdate, normalizeAndDiff, setNestedValue } from "../src/data-updating.ts";
+import { unwrapOneLevelAndBindtoDoc, sendValueToBinding } from "../src/recipe-binding.ts";
+import { followAliases } from "../src/link-resolution.ts";
+import { isEqualCellLink } from "../src/type-utils.ts";
+import { extractDefaultValues, mergeObjects } from "../src/runner.ts";
 import { Runtime } from "../src/runtime.ts";
 import { CellLink, isCellLink } from "../src/cell.ts";
 import { type ReactivityLog } from "../src/scheduler.ts";
